Add the week's daily highs and lows, and widen the page mask to hold them
Three more pages: seven days side by side, each a bar from that day's
low to its high, with the day's initial underneath and the week's own
range in the title. Unlike every other chart here the seven bars share
one scale, which is the point -- the question is "was yesterday colder
than today", and days each scaled to themselves would draw seven
identical bars. A day with no readings is dotted rather than drawn at
floor height, so a board switched on yesterday does not invent a very
cold week.

Adding the thirty-third page broke the page mask, and the menu's own
test caught it: PAGE_BIT was `1u << page` on a 32-bit unsigned, so page
32 was undefined behaviour that in practice wrapped to bit 0 and aliased
the clock. A board with three pages offered a fourth it did not have.

The mask is 64 bits now, with a static assertion that the page count
still fits it -- adding an ordinary page should not be able to do this
silently.

The same widening then hid a second instance of itself. `imu_pages`
stayed a plain unsigned, so `available &= ~imu_pages` zero-extended the
complement and cleared every page above the thirty-second: the pressure
week page simply never appeared in the rotation, with nothing from the
compiler to say why. There is a page_mask_t now and a note saying why it
is spelled out, because the failure mode is silent in both directions --
a page that appears and should not, and one that should and does not.

+++ host_tests/test_views.c
@@ -520,7 +520,7 @@ int main(int argc, char **argv)
     /* The menu, and whether every tile can be hit. */
     {
         pages_t pg;
-        unsigned all = 0;
+        uint64_t all = 0;
         for (int i = 0; i < PAGE_COUNT; i++) all |= PAGE_BIT(i);
         pages_init(&pg, all);
         pages_show(&pg, PAGE_MENU, 1000);
